demo/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
jsonschema==4.20.0
python-dateutil==2.8.2
loguru==0.7.2
azure-cosmos==4.5.1
pytest==7.4.3
pytest-flask==1.3.0
